Report
Given the provided data, what are three conclusions we can draw about Kickstarter campaigns?
Looking at the line chart, we can say that when it started in January, it was making a lot of progress, it starts going down in mid-September. Also, there is a lots of business ideas that failed or cancelled because of lack of pledged. And also, the successful rate is way better than the failure rates.
What are some limitations of this dataset?
I couldn’t think of any at this point, but I’m pretty sure that are some. 
What are some other possible tables and/or graphs that we could create?
There are so many graphs or charts we could of use , but the line chart is more descriptive.